refactor: keep the preset run receipt types out of the contract file namespace
